implemented warehouse and exception

+++ task_3/src/main/java/com/shichko/shape/repository/impl/IdSpecification.java
@@ -5,9 +5,9 @@ import com.shichko.shape.repository.Specification;
 
 public class IdSpecification implements Specification {
 
-    private int id;
+    private long id;
 
-    public IdSpecification(int id) {
+    public IdSpecification(long id) {
         this.id = id;
     }
